fixed planificator modal & moved it to modals folder
