+++ components/modals/planificator.vue
